requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow
scikit-learn
Pillow
streamlit-drawable-canvas